worker/requirements.txt
# Phase 1 Processing Worker Dependencies

# Celery and Redis
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# Geospatial processing
rasterio==1.3.9
pyproj==3.6.1
Shapely==2.0.2
Fiona==1.9.5

# Cloud storage
boto3==1.34.34
minio==7.2.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Image processing
Pillow==10.1.0
numpy==1.24.4

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Progress tracking
tqdm==4.66.1